LayoutTests/imported/w3c/web-platform-tests/css/selectors/focus-visible-script-focus-015-expected.txt

This test checks that a script focus after blur after a keyboard input does match :focus-visible.

Click on the input below and type some letter.
If the element that says "Focused" has a red background then the test result is FAILURE, if it has a green outline then the test result is SUCCESS.

Focused

PASS ":focus-visible" should be a valid selector
FAIL Script focus after blur after keyboard input does match :focus-visible assert_equals: outlineColor for DIV#target should be green expected "rgb(0, 128, 0)" but got "rgba(46, 52, 54, 0.59)"
